+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Please choose versions by [Semantic Versioning](http://semver.org/).
 * MINOR version when you add functionality in a backwards-compatible manner, and
 * PATCH version when you make backwards-compatible bug fixes.
 
+## v1.13.1
+
+- fix TimePtr on nil Date, DateTime or UnixTime
+- add AddTime to Date, DateTime or UnixTime
+
 ## v1.13.0
 
 - remove vendor